+++ src/bl64-py-setup.bash
@@ -93,6 +93,7 @@ function _bl64_py_set_command() {
     ${BL64_OS_FD}-33.* | ${BL64_OS_FD}-34.*) BL64_PY_CMD_PYTHON39='/usr/bin/python3.9' ;;
     ${BL64_OS_FD}-35.* | ${BL64_OS_FD}-36.*) BL64_PY_CMD_PYTHON310='/usr/bin/python3.10' ;;
     ${BL64_OS_FD}-37.* | ${BL64_OS_FD}-38.*) BL64_PY_CMD_PYTHON311='/usr/bin/python3.11' ;;
+    ${BL64_OS_FD}-39.* ) BL64_PY_CMD_PYTHON312='/usr/bin/python3.12' ;;
     ${BL64_OS_DEB}-9.*) BL64_PY_CMD_PYTHON35='/usr/bin/python3.5' ;;
     ${BL64_OS_DEB}-10.*) BL64_PY_CMD_PYTHON37='/usr/bin/python3.7' ;;
     ${BL64_OS_DEB}-11.*) BL64_PY_CMD_PYTHON39='/usr/bin/python3.9' ;;
@@ -123,7 +124,10 @@ function _bl64_py_set_command() {
     esac
 
     # Select best match for default python3
-    if [[ -x "$BL64_PY_CMD_PYTHON311" ]]; then
+    if [[ -x "$BL64_PY_CMD_PYTHON312" ]]; then
+      BL64_PY_CMD_PYTHON3="$BL64_PY_CMD_PYTHON312"
+      BL64_PY_VERSION_PYTHON3='3.12'
+    elif [[ -x "$BL64_PY_CMD_PYTHON311" ]]; then
       BL64_PY_CMD_PYTHON3="$BL64_PY_CMD_PYTHON311"
       BL64_PY_VERSION_PYTHON3='3.11'
     elif [[ -x "$BL64_PY_CMD_PYTHON310" ]]; then